Add notext linker option due to position-dependent assembly code
The assembly code uses position-dependent data access, which requires
text relocations when the test program is built as a position-independent
executable (PIE). This option indicates that text relocations will be used.

In the future, text relocations will be eliminated by placing temporary data
on the stack instead of a separate data segment.